requirements.txt
Flask
flask-cors
python-dotenv
PyPDF2
google-generativeai
gunicorn
